+++ CentralStation/Api/StopNetworks.cs
@@ -1,10 +1,10 @@
 using System;
 
-namespace Pathoschild.Stardew.CentralStation.Framework;
+namespace Pathoschild.Stardew.CentralStation;
 
 /// <summary>The interconnected networks that join all the stops of a given type.</summary>
 [Flags]
-internal enum StopNetworks
+public enum StopNetworks
 {
     /// <summary>The stop can be reached by train.</summary>
     Train = 1,